+++ w2d1/BaseERC20WithOpenzeppelin.sol
@@ -4,10 +4,7 @@ pragma solidity ^0.8.4;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 interface TokenRecipient {
-    function tokensReceived(
-        address sender,
-        uint amount
-    ) external returns (bool);
+    function tokensReceived(address sender, uint256 amount) external returns (bool);
 }
 
 contract BaseERC20WithOpenzeppelin is ERC20 {
