+++ w2d4/src/Bank.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.25;
 
 contract Bank {
-    mapping(address => uint) public balanceOf;
+    mapping(address => uint256) public balanceOf;
 
-    event Deposit(address indexed user, uint amount);
+    event Deposit(address indexed user, uint256 amount);
 
     function depositETH() external payable {
         require(msg.value > 0, "Deposit amount must be greater than 0");
