+++ w2d4/src/NoahNFTMarket.sol
@@ -6,13 +6,13 @@ import "@openzeppelin/contracts/token/ERC721/IERC721Receiver.sol";
 import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
 
 contract NoahNFTMarket is IERC721Receiver {
-    mapping(uint => uint) public tokenIdPrice;
-    mapping(uint => address) public tokenSeller;
+    mapping(uint256 => uint256) public tokenIdPrice;
+    mapping(uint256 => address) public tokenSeller;
     address public immutable token;
     address public immutable nftToken;
 
-    event List(uint indexed tokenId, uint price);
-    event Purchase(uint indexed tokenId, address buyer);
+    event List(uint256 indexed tokenId, uint256 price);
+    event Purchase(uint256 indexed tokenId, address buyer);
 
     // 在部署合约的时候将 token合约地址 和 nftToken地址 传入
     constructor(address _token, address _nftToken) {
@@ -21,32 +21,20 @@ contract NoahNFTMarket is IERC721Receiver {
     }
 
     // approve(address to, uint256 tokenId) first
-    function list(uint tokenID, uint amount) public {
-        IERC721(nftToken).safeTransferFrom(
-            msg.sender,
-            address(this),
-            tokenID,
-            ""
-        );
+    function list(uint256 tokenID, uint256 amount) public {
+        IERC721(nftToken).safeTransferFrom(msg.sender, address(this), tokenID, "");
         tokenIdPrice[tokenID] = amount;
         tokenSeller[tokenID] = msg.sender;
         emit List(tokenID, amount);
     }
 
-    function purchase(uint tokenId) external {
-        require(
-            IERC721(nftToken).ownerOf(tokenId) == address(this),
-            "aleady selled"
-        );
+    function purchase(uint256 tokenId) external {
+        require(IERC721(nftToken).ownerOf(tokenId) == address(this), "aleady selled");
 
         // require(IERC721(nftToken).ownerOf(tokenId) == address(this)msg.sender != tokenSeller[tokenId], "aleady selled");
 
         // 从 msg.sender 搞钱 给 tokenSeller[tokenId] (卖家)
-        IERC20(token).transferFrom(
-            msg.sender,
-            tokenSeller[tokenId],
-            tokenIdPrice[tokenId]
-        );
+        IERC20(token).transferFrom(msg.sender, tokenSeller[tokenId], tokenIdPrice[tokenId]);
         // 从我(当前合约) 这里搞 NFT Token(by tokenID) 给 msg.sender
         IERC721(nftToken).safeTransferFrom(address(this), msg.sender, tokenId);
 
@@ -54,12 +42,7 @@ contract NoahNFTMarket is IERC721Receiver {
     }
 
     // 实现{IERC721Receiver}的onERC721Received，能够接收ERC721代币
-    function onERC721Received(
-        address,
-        address,
-        uint256,
-        bytes calldata
-    ) external pure returns (bytes4) {
+    function onERC721Received(address, address, uint256, bytes calldata) external pure returns (bytes4) {
         return this.onERC721Received.selector;
     }
 }
